+++ src/cli/mc.ts
@@ -12,6 +12,7 @@ import dataExchangeCommand from "./commands/data-exchange";
 import dataLakeCommand from "./commands/data-lake";
 import deleteAssetCommand from "./commands/delete-asset";
 import deleteFileCommand from "./commands/delete-file";
+import deliveryJobsCommand from "./commands/delivery-jobs";
 import downloadEventsCommand from "./commands/download-events";
 import eventTypesCommand from "./commands/event-types";
 import eventsCommand from "./commands/events";
@@ -75,7 +76,6 @@ import signalValidationCommand from "./commands/signal-validation";
 import spectrumAnalysisCommand from "./commands/spectrum-analysis";
 import subtenantCommand from "./commands/subtenant";
 import tenantCommand from "./commands/tenant";
-import tokenExchangeCommand from "./commands/token-exchange";
 import trendPredictionCommand from "./commands/trend-prediction";
 
 const program = new Command();
@@ -108,6 +108,8 @@ iotBulkDirCommand(program);
 iotBulkRunCommand(program);
 iotCheckBulkComand(program);
 iotDownloadBulkCommand(program);
+
+deliveryJobsCommand(program);
 
 // * assets files and event handling commands
 assetModelLockCommand(program);
@@ -149,7 +151,7 @@ identityCommand(program);
 policyCommand(program);
 serviceCredentialsCommand(program);
 serviceTokenCommand(program);
-tokenExchangeCommand(program);
+// tokenExchangeCommand(program);
 
 // * data lake
 dataLakeCommand(program);